--- a/client/src/App.css
+++ b/client/src/App.css
@@ -198,6 +198,9 @@ label { user-select: none; }
 .full-width { width: 100% !important; }
 .no-border, .no-border tr td  { border: none !important; }
 .no-outline { outline: none !important; }
+.oauth-box span#disconnect { display: none !important; }
+.oauth-box:hover span#connect { display: none !important; }
+.oauth-box:hover span#disconnect { display: inline !important; }
 .disabled, .disabled:hover {
     cursor: auto;
     opacity: 0.7;