--- a/server/lib/crypto.js
+++ b/server/lib/crypto.js
@@ -1,18 +1,19 @@
 const crypto = require('crypto');
 const algorithm = 'aes-256-cbc';
-const secretKey1 = process.env.SECRET_KEY_1;
-const secretKey2 = process.env.SECRET_KEY_2;
-const secretKey3 = process.env.SECRET_KEY_3;
+const secretKey1 = process.env.SECRET_KEY_1; // Account Session and Todo Data
+const secretKey2 = process.env.SECRET_KEY_2; // OTP Verification Code
+const secretKey3 = process.env.SECRET_KEY_3; // Forgot Password Credentials
+const secretKey4 = process.env.SECRET_KEY_4; // Backup Codes Credentials
 const iv = crypto.randomBytes(16);
 
 const encrypt = (text, method) => {
-    let cipher = crypto.createCipheriv(algorithm, Buffer.from(String(method) === '1' ? secretKey1 : String(method) === '2' ? secretKey2 : secretKey3), iv);
+    let cipher = crypto.createCipheriv(algorithm, Buffer.from(String(method) === '1' ? secretKey1 : String(method) === '2' ? secretKey2 : String(method) === '3' ? secretKey3 : secretKey4), iv);
     let encrypted = Buffer.concat([cipher.update(text), cipher.final()]);
     return { iv: iv.toString('hex'), data: encrypted.toString('hex') };
 }
-   
+
 const decrypt = (hash, method) => {
-    let decipher = crypto.createDecipheriv(algorithm, Buffer.from(String(method) === '1' ? secretKey1 : String(method) === '2' ? secretKey2 : secretKey3), Buffer.from(hash.iv, 'hex'));
+    let decipher = crypto.createDecipheriv(algorithm, Buffer.from(String(method) === '1' ? secretKey1 : String(method) === '2' ? secretKey2 : String(method) === '3' ? secretKey3 : secretKey4), Buffer.from(hash.iv, 'hex'));
     let decrypted = Buffer.concat([decipher.update(Buffer.from(hash.data, 'hex')), decipher.final()]);;
     return decrypted.toString();
 }