**What's Changed in v0.5.2:**
1. Added Disconnect `Account` with `OAuth` Feature (#146) (14fb4b6d95ac18973ad7d01e465bf7965c7ec85c) (6cc1f3c8892811d9b19c5dcba33be62210943563)
2. Added `js-file-download` Dependency (#142) (bb644d1b486d9cbdb2518827145e27d75172763e)
3. Added Download Backup Codes Feature (#139) (dd326b36c091961866bbd5e48932a60aab717af5)
4. Added `Encryption` Method for Backup Codes (#140) (76c0d969b9161aa6c2c0f598adf8271ddc97e460)
5. Upgrade Dependencies to the Latest Version (068a9473a325380b5efb02fd6c1e6b9542f7e709):
    * `date-fns` dependency to `2.21.3`
    * `web-vitals` dependency to `1.1.2`
    * `dotenv` dependency to `9.0.1`
6. Updated `Rate Limiter` Settings on `Login` and `Register` `API` (97fcb8f448085e3872605021b6a9adb26f407445)
7. Updated `Passport` `Security` and Code Quality (f1831806a43abc4e6f97cedfccd12bb229ffcae4)
8. Optimized `Code` Quality in `OAuth` Route (249c40f18413673df89ce4e1c9afbab59d3072de)
9. Optimized `Code` Quality in `Todo` Route (2fbcf5bcb384f72ebfc2afb5e42a98aa314c1eaa)
10. Optimized Speed in User `API` (04e8e134a425de2c64cced889dc205a4a045998c)
11. Optimized `Server` `Models` (122045478b5f6388e1b98dc69d75d22939c22abf)
12. Optimized `Server` `API` (a9682d9267c939b50ed76c8f2aa5a9a45a2e4c1e)(0e19880dec40fdf5ea8fbc844a25087cf54efc43) (341d64ae9cecdd8f13c2018b66ae7bc2cbd44bdf) 
13. Improved `Security` (04e8e134a425de2c64cced889dc205a4a045998c) (2fbcf5bcb384f72ebfc2afb5e42a98aa314c1eaa) (249c40f18413673df89ce4e1c9afbab59d3072de)
14. Resolved Issues of:
    * Received Missing Credentials Message (#138)
    * `Security` of `MFA` (2c95f6e2081a2db3ff22c187fa2303e48e72a4ea)
    * File Still Exists in `Server` After Backup Codes are Downloaded (#141) (d3e2d4a8ab96a116f1a2c5fd279ef1c8c38c14c5)
    * `Server` Crashes while Responding `504` Error Code (#143) (4e139b80b2c428a5b3a40eb2d8267b9f4559ea04)
    * `Unverified` Users are Able to Send `OTP` (#144) (700d5d2d842c0a63eafb593ba87d5388b077d334)
    * `OTP` Code Still Exists while Operations Succeed (#147) (9ccab7b067388aaae801251351d5d86cb0b7d905)
    * Fixed Minor Bugs

Details About Changes in v0.5.1:
| Client | `build\static\js\2.6ca8d93c.chunk.js` | `build\static\js\main.2591d28a.chunk.js` | `build\static\css\main.030489ff.chunk.css` | `build\static\js\3.668acfa5.chunk.js` | `build\static\js\runtime-main.3cbd5898.js`
------ | --------------------------------------- | ------------------------------------------- | ---------------------------------------------- | -------------------------------------------------- | ---------------------------------------------- |
| v0.5.2 | 152.63 kB (+200 B) | 30.03 kB (+220 B) | 4.75 kB (+20 B) | 1.62 kB (+60 B) | 1.16 kB |
| v0.5.1 | 152.43 kB (+165 B) | 29.81 kB (+1.47 kB) | 4.73 kB (+193 B) | 1.56 kB | 1.16 kB |
| v0.5.0 | 152.27 kB (+3.16 kB) | 28.34 kB (+428 B) | 4.54 kB (+37 B) | 1.56 kB | 1.16 kB |
| v0.4.9 | 149.11 kB (+159 B) | 27.92 kB (+115 B) | 4.5 kB (+1 B) | 1.56 kB | 1.16 kB |
| v0.4.8 | 148.95 kB (+42.58 kB) | 27.81 kB (+2.2 kB) | 4.5 kB (+224 B) | 1.56 kB (+2 B) | 1.16 kB (+2 B) |

| Server | Size |
| ------- | ---- |
| v0.5.2 | 216.32 kB (- 2.86 kB) |
| v0.5.1 | 219.18 kB (+30 B) |
| v0.5.0 | 219.15 kB (+7.2 kB) |
| v0.4.9 | 211.95 kB (+95 B) |
| v0.4.8 | 211 kB (+21 kB) |

For more information about changes in this release, please visit our [project roadmaps v0.5.2](https://github.com/stanleyowen/todo-application/projects/32)